更新弹幕姬SDK

+++ Re_TTSCat/Data/UserGift.cs
@@ -5,13 +5,13 @@ namespace Re_TTSCat.Data
     public class UserGift
     {
         public string User;
-        public int UserId;
+        public long UserId;
         public string Gift;
         public int Qty;
         public DateTimeOffset TargetTime;
 
         public UserGift() => TargetTime = DateTimeOffset.Now + TimeSpan.FromSeconds(Vars.CurrentConf.GiftsThrottleDuration);
-        public UserGift(string user, int userId, string gift, int qty) : this()
+        public UserGift(string user, long userId, string gift, int qty) : this()
         {
             User = user;
             UserId = userId;
